Introduce a Gradle platform plugin
The "gradle-platform" plugin lets the user create a component type
which corresponds to a "Gradle platform". A Gradle platform is a
set of recommendations which are eventually used to suggest plugin
versions or dependency versions.

Those basically correspond to what is possible to declare in a
settings file or a dependencies.toml file.

As of now, it's only possible to produce such a platform, but
NOT to consume it.

Consumption is going to be implemented in a subsequent commit.