Provide a way to build a Gradle platform from a `Configuration`
This is done as a convenience: all dependencies declared in the
`gradlePlatform` configuration (as well as the constraints) will
be assigned an alias corresponding to their module name.

If there's a name clash, for example if multiple modules have the
same name but different groups, then the build fails, unless an
explicit alias is configured.

This is done so that we can make it convenient to publish a Gradle
platform from a Java platform, which is going to be done in a
different commit.