Enhance the Java Platform plugin to produce a Gradle platform
This commit wires in the Java platform plugin so that if a project
also applies the "Gradle Platform" plugin, then the Java platform
publication also publishes a Gradle platform.

The model is automatically inherited from the runtime configuration
of the Java platform.

This makes it extremely convenient to publish a platform both as
a BOM and a "Gradle plaform".