'included build name' and 'root build root project name' are not in conflict (#15012)
This removes a restriction which was introduced in b5c1d4e

It is unclear why this was introduced, as there is no integration test
showing a problematic case. The root build's root project name is not
used for an (identity) path. The path of that build is always ':'.
Hence there should be no need for this restriction.