Add support for importing a Gradle platform from settings
This effectively closes the loop: the settings can be used
to declare a set of "recommendations" for dependency versions,
a catalog to read from (either plugins or dependencies).

Then the `gradle-platform` plugin makes it possible to build
such recommendations. It is also possible to use the java
platform plugin as a convenience for this.

Once the platform is published, it can be consumed from
the settings, so users have the choice to consume either
from a local file or a published platform.

The DSL in settings makes it convenient to actually use
different extensions for different platforms, so it's
possible to separate concerns (libraries, test libraries,
plugins) as much as needed.