Support cyclic build includes in IDE import
Although it is possible since a long time to have cycles through
'includeBuild(...)' statements IDE support never supported this.
There were several places in the code assuming there won't be cycles
which ran into infinite recursions. These places are fixed and
the missing test coverage is added.

Since this concerns both idea/eclipse plugin code and TAPI code,
this fix requires both the running Gradle and the TAPI to be on the
latest version.