CodeQL: Analyse the merge commit in pull requests
We have recently changed our default workflow to analyse the merge commit for pull requests. Among other effects this makes it easier for us to correctly identify the best commit of the base branch to compare against. This should help to reduce the incidence of errors that you reported recently.

Signed-off-by: Robin Neatherway <rneatherway@github.com>