Make it possible to filter what is imported from a catalog
When building a dependencies model, Gradle issues warnings whenever
a dependency is overriden, so that it's not accidentally overridden.
In order to be able to compose catalogs from other catalogs and
avoid such potential conflicts, the `from` method now accepts an
optional filter which lets the user declare what to include or
exclude.