Remove unused code for handling public inline functions in Kotlin compile avoidance
Compile avoidance is not supported for public inline functions - we fall back to using the full hash of the jar upon encountering a public inline function. This change removes previously added code that attempted to make public inline functions part of the ABI for compile avoidance. This code is no longer used.