Add “mergeReruns” option to JUnit XML report generation
When enabled, the XML output will be very similar to Apache Maven's surefire plugin when enabling reruns.
Most notably, if a test fails but is then retried and succeeds, its failures will be recorded as <flakyFailure> instead of <failure>.
This can be important for build tooling that uses this XML to understand test results, and where distinguishing such passed-on-retry outcomes is important.
This is the case for the Jenkins CI server and its Flaky Test Handler plugin.

This does not add any retry/rerun functionality.
It just changes the reporting if such a mechanism is separately in place.