Fix dependency verification in init scripts
If an init script was using external dependencies and that
those dependencies were not present in the dependency verification
file, the build would fail with an error message saying that the
GradleProperties couldn't be loaded.

This commit changes the behavior so that we perform verification and
report a verification error message instead.

Fixes #14711