Provide more context when competing constraints can't be satisfied
The bug report #15009 was difficult to diagnose because the error
message didn't contain enough context to make sense of what we
were seeing. In practice, a transitive dependency was using an
enforced platform, which causes all of its dependencies to be
forced. Therefore, the mistmatch between the message, which was
saying that a strict constraint couldn't be satisfied, and what
we saw, which is "normal" versions, didn't make any sense.

This commit adds the selected variant to the error message, so
that at least we can see that in a particular path, a specific
variant is selected. In this case, seeing "enforcedPlatform"
would have been a red herring.