Only expire on full metaspace if GC rate is high
The metaspace expiration strategy expired daemons too eagerly.
It only looked at the space, but not whether the GC was busy.
A full metaspace alone is no big problem and happens very easily
with Gradle due to the many soft reference caches. These soft refs
are only cleaned up when the JVM runs out of memory.

The strategy will now only expire the daemon if metaspace is full
and the GC fails to free up memory (i.e. there is a real leak).

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>