Only use the identity name of an included build if there is a conflict
This is a follow up to #14671 to preserver the previous behaviour for
projects that already work in the IDE.

Theoretically, it should not matter what the name of the IDE
project/module is as long as it is unique for each one.
Still, currently some of the IDEA Sync code depends on the previous
behaviour to setup dependencies between included builds correctly.

#14671 changed the behaviour to always use the 'identity name' if
available. We now revert to the previous behaviour of using the
normal projects name. Instead, when we reach a point in the
de-duplicator where we still have conflicts but cannot find a solution
by adding more prefixes, we then change to use the identity name
instead of the project name.