Fix addition and deletion of generated types
The previous fix for #13767 was incorrect because it assumed
that if a type didn't have any source, then it shouldn't be
reprocessed. In practice, we need to keep track of all types
which contribute to aggregation, but when we compute the list
of types to be reprocessed, instead of adding the type directly,
we need to figure out what source file actually caused the
generation of the type to process.

For example, imagine an isolating processor P1, which, from
a source file A, generates a class AHelper, and a source file
B, generates BHelper. These generated classes are in turn
annotated with an annotation which is handled by processor P2,
which is aggregating and generates a service registry.

Then, if A is deleted, we need to reprocess B, that is to say
the type which generated BHelper, not BHelper itself.

This fixes #15097 as well as properly fixes #13767