Improve diagnostic logging on work unit timeout (#15274)
Previously, there was no indication in the logging output that a work unit (e.g. task) was being stopped due to timeout.
This made it hard to diagnose when the work unit failed to respond to the stop, as the build would appear to just hang with no indication as to why.

Now, a message is always logged when the timeout occurs and stop is requested.
If the work unit takes longer than 3s to stop, a message will be logged indicating that it has not yet stopped.
This message will be logged every 3s until the work unit stops.
If this occurs, an extra message will be logged when the work unit finally stops.

Fixes #14960.