Validate that publications do not depend on enforced platform
This commit adds validation of attributes when generating Gradle
Module Metadata. We do this at generation of metadata so that it's
not too late in case of publishing.

There we can warn that a publication contains something suspicious.
If the user knows what they are doing, they can disable the validation
by adding a key to the `suppressedValidationErrors` set on the
metadata generation task.

See #15009