Continue to interrupt timed out work unit until it stops
Previously, it was possible for a work unit to clear the received interrupt and continue working, which is not the intended outcome.

While it's still possible for a work unit to be written in a way that intentionally repeatedly ignores the interrupts,
this change makes it less likely that a work unit fails to stop because an InterruptedException was caught and incorrectly ignored.