Fix edge case in local consistency resolution
The previous implementation of local consistency resolution added
explicit constraints to the configuration being resolved, just
before resolution happens. It works fine if no other configuration
actually extends from this configuration.

In case another resolvable configuration extends from this one *and*
that it is resolved before, then there are edge cases where depending
on the order of resolution, a configuration may see dependency
constraints being added _after_ it was actually included in dependency
resolution.

For example, let's say we resolve `myClasspath` first, and that `myClasspath`
extends from `compileClasspath`, which itself is configured to be
consistent with `runtimeClasspath`.

Then triggering the resolution of `myClasspath` will lock the hierarchy,
which includes `compileClasspath`. When we will try to resolve `compileClasspath`
later, then we will resolve `runtimeClasspath` and try to add constraints
to `compileClasspath`. But because it was already used in the config
hierarchy of `myClasspath`, it's been locked and addition of constraints
fail.

In order to avoid the problem, because what we care about is actually
injecting constraints _only for resolution_, the new implementation works
a bit like dependency locking, by injecting constraints which are only
visible by the graph builder. That is to say that those constrainsts are
no longer "real" constraints added to the configuration.

It changes nothing in terms of reporting, only we make sure that constraints
which are only aimed at resolution time are used at resolution time. In
other words, the fact that a resolvable configuration also acts as a "bridge"
to get dependencies when creating a configuration is restored.