Write `BuildServiceProvider` instances only once and in the correct cache
Before this change, `BuildServiceProvider` instances would be written to the
configuration cache in the context of the build in which they are
referenced, no matter which build they were registered with. That has
implications when build services are registered as build event listeners since
event listener subscriptions are stored in the context of the root build.

After this change, `BuildServiceProvider` instances found while writing event
listener subscriptions to the cache are simply marked as used and then later
written to the cache in the context of the correct build.