Remove optimistic caching of the component metadata handler
When reworking the creation of the component metadata handler to
support settings level component metadata rules, we assumed that
rules were added _before the first resolution in a project_.

This assumption doesn't hold: it is possible, today, to resolve
a configuration, _then_ add rules. While in theory we should
disallow this, a popular plugin, the "Nebula Resolution Rules"
plugin, relies on this behavior, because it uses the dependency
resolution engine itself to download rules which are then
added to the project.

Fixes #15312