Fix display names not being propagated to the TAPI events
This commit is a first step in fixing some of the test reporting
issues described in #5975. It will now propagate the display
names of test classes and methods to the TAPI generated events.

Note that this solves the TAPI events, but it will typically not
fix what you see in IntelliJ IDEA because they are actually not
consuming the "displayName" of test descriptors.