Rework build logic build inclusion
When evaluating pluginManagement in settings, create included build definitions to be registered later instead of calling the actual build inclusion logic in the build state registry. Now build state registry keeps track of build definitions that might need to be included early as determined by EarlyCompositeBuildPluginResolverContributor. If EarlyCompositeBuildPluginResolverContributor is not invoked, then such builds are included at the same time as regular included builds.
This fixes the case where builds would be included too early when executed with configuration cache.