Fix Kotlin DSL sync when failing on project repositories
When using the central repositories declaration and the Kotlin DSL,
IDE sync could fail if the repositories mode was set to failing on
project repositories.

The reason is that the Kotlin DSL sources provider was adding a
repository transparently, to resolve the Gradle sources, then removed
it silently.

However, this behavior is not allowed when failing on project
repositories.

This commit fixes the problem by making sure the Kotlin DSL sources
provider works in isolation, not leaking any repository to the
project. To do this, we introduced an _internal_ detached resolver,
which is basically an isolated dependency resolution engine, with
its own repositories, configurations, dependencies, ...

The goal of this PR is _not_ to provide a public API for this, but
use a pragmatic solution for fixing the problem mentioned above.