Isolate daemons for Kotlin/Android smoke tests
We test so many combinations of these plugins that they will eventually
use up all Metaspace. They can't be garbage collected because the Kotlin
plugin leaves a connection to its compiler daemon open, which prevents
the Kotlin plugin classloader from ever being garbage collected. This
is not an issue in most cases, since users will rarely switch back
and forth between as many versions as we are testing here.

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>