Reuse matching toolchains
By reusing matches for toolchains given a specific filter, we can avoid
the matching if the provider is queried multiple times
(e.g. `isPresent`).
Further, this allows to short-circuit looking up a provisioned JDK.
If no jdk is found and a new one is provisioned, it is returned as a
toolchain directly from the provisioning service. If you query the same
provider again, we don't detect the new toolchain via the regular means
(as they're cached per daemon) and thus incur the cost of looking up
the archive and trying to reprovision. By reusing the query match, this
can be avoided.

See issue #15325