Avoid exercising the executable
To determine the java version, we used to accept a java command. This
can but must not end with an executable file extension (see `PATHEXT`
environment variable). So `/some/path/java` is valid on Windows even
though when executing, it would use `java.exe`. Thus, we need to avoid
checking for file existence for the given executable.