Improve error message if settings repositories are ignored
In case a project declares repositories but that the repositories
mode is set to the default, which is to use project repositories,
then if resolution fails it can be hard to figure out why.

This commit improves the situation by checking if a resolution
failure is due to a missing module. If this is the case and that
resolution used project repositories instead of settings repositories,
we complete the error message with a hint that it could be the
problem.

Closes #15336