Provide access to version catalogs from plugins
Version catalogs are dynamically generated from definitions in settings.
This means that when we build a plugin, we actually don't have access to
the catalogs, because we don't know if there will be one or not.

In particular, it's not because a project declares a catalog that the
catalog is visible in `buildSrc` plugins.

This commit exposes an API which provides access to catalogs using
"unsafe" APIs. The user is expected to check that aliases, bundles, etc
are available.

Fixes #15382