Remove `compile` and `runtime` configurations
This might be called Grapocalypse.

This commit removes the "compile" and "runtime" configurations,
which were soft deprecated for years and hard deprecated in the
6.x line of Gradle.

The replacement for `runtime` is straightforward: go with `runtimeOnly`.
The replacement for `compile` requires a bit more thinking, depending
on the use case: the natural replacement is to use the `api` configuration.
However, older builds may use the `java` plugin in which the `api`
configuration is not available. In this case, the first step is to migrate
to the `java-library` plugin first, then use the `api` configuration.
However, in a lot of cases, the actual configuration to use is not `api`
but `implementation`: `api` should only be used when the dependencies
are part of the API of a class (method signatures, public fields, ...).