`ktlintKotlinScriptCheck` should only check the build files
And not all the *.kts files in the project directory. Checking all
those would mean that the check can't run in parallel with any other
task which produces anything in the project. It also made the task
fail, since we detected missing dependencies between this task
and any other task in the same build.