Restore the `MavenPlugin` class
A number of popular plugins are not ready for 7.0, mainly because
they use the `plugins.withType(MavenPlugin)` pattern to react on
the application of the legacy maven plugin. By removing the type
in 7, we actually break those plugins which wouldn't load anymore,
even if the maven plugin is not applied.

This commit "fixes" this by re-introducing the `MavenPlugin` class,
but restricting its scope (making the class abstract).

Therefore, plugins won't be able to compile against 7.0, but will
still be able to _run_ on 7.0.