Fix excludes being ignored when using `available-at`
This commit fixes a regression introduced by a bugfix
about `available-at` (see 4be5952141fc64f0f1f754807730b1b18bfb32b5).

Basically the problem is that if an exclude from an incoming edge
is supposed to exclude the target of an available-at variant, then
the exclusion would be ignored. In fact, instead of allowing all
dependencies like we did in the previous fix, we should only allow
non excluded variants, which corresponds to the union of the
"transitive exclusions" of all the incoming edges to a node.