Automatically add required exclusion for JDK 9+ & includeNoLocationClasses
Fixes #5184

When includeNoLocationClasses is set to true, and a JVM version 9 or above is used, the JaCoCo plugin fails with an exception. While using includeNoLocationClasses is uncommon it is used by a number of folks who currently have to all add a manual workaround to their builds. This change moves the fix into the plugin code so folk don't need to add the manual workaround.

Signed-off-by: Al Sutton <al.sutton@alsutton.com>