Workaround setup/Before ordering in Spock 2
Spock 2 executes Before annotated JUnit methods after its setup() methods which used to be the other way around.
This breaks the tests that made use of this specific fixture in their setup() methods where it was not yet initialized.
This change attempts to provide the least invasive workaround for the time being.