Simplify tooling-api test - remove always-true condition
This test is constrained to always run with target Gradle versions >=2.13 - there is no point in additionally checking whether the target version is greater than 1.12 - it always is. This change also makes it easier to work around ordering issues with when the targetDist is set by the testing framework when migrating to Spock 2.