Replace JUnit4 Rule with explicit Spock setup/cleanup
While JUnit4 rules work with Spock2 using spock-junit4 library, that's only to ease the migration.
In this particular case, replacing rules with explicit setup/cleanup works around the ordering of when the rule in question is created and when the testing framework sets the targetDist for the GradleDistribution under test when using Spock2.