Fix worker classpath vs. module path setup (#15704)
Gradle attempted to run a worker as module in too many cases. Having
the 'infer' option enabled is not enough. We also actually need a
module to run on the module path. The fix that was done in #15677 is not
correct. With this commit, the condition is now as follows:

- Only run the worker itself as module when the application has
  a module path
- Only in that case, the implementation module path of the worker
  itself is taken into account. Otherwise everything is done on
  the classpath.

The issue was discovered by a test running against the play framework
plugins with a very unique worker setup. The issue should have also
been revealed by all tests for the JUnit5 test runners, because JUnit5
actually has a specific module path when running as module. However,
this was hidden by specifically setting 'inferModulePath=false' in the
case of test running under a specific condition. This special handling
is now removed.