Instantiate test repos in test setup
So that they are always instantiated with the correct and current roots in the filesystem.
Those depend on the temporary directory in use and it changes with each test iterations.
Having lazy semantics with getters here, causes multi-version tests to use repository roots from previous test iterations in multi version tests. This breaks the cleanup rule and results in test content not being cleaned up properly.