Ignores missing dependency if there is an explicit ordering
For some use-cases there are cyclic dependencies with respect to input
and output location of tasks. For example the spotless apply task depends
on the output of the spotless analyze task, but also modifies the inputs
(the sources) of the spotless analyze task. We don't have a better solution
for this use-case, yet, so we don't want to emit a deprecation warning.

So as long as there is an order relation in any direction, we do not emit
a warning about missing dependencies between tasks.

See #15616.