Make it an error to not handle duplicates explicitly
Previous Gradle releases used a default 'include' strategy for
copy specs: if a duplicate was found, then Gradle would silently
include the duplicate, potentially overwriting the previous entry.

In Gradle 5, we introduced a warning whenever duplicates are found
and that no explicit deduplication strategy is set. With Gradle 7,
this now becomes an error.

Closes #15594