Take filtering into account for missing dependency detection
This allows for the use-case to zip up the root project
directory, as long as no outputs of other tasks are not
included in the zip.

Note that filters for outputs are not taken into account,
only filters for inputs. This is because for outputs,
even when the output is a filtered file tree, the task
still may produce files anywhere in the output directory.