Load only inner classes when discovering JUnit platform tests
Those classes are loaded by Gradle in an isolated classloader to assert whether they should be passed to JUnit engine for further discovery. Because the logic only applies to inner classes, we can tell whether the class is an inner one by looking for a $ in its name - this way we do not have to classload classes that are not inner classes.