Rework tooling cross-version tests
Split the cross version tasks so that they always have a single iteration/execution - separate task for "TAPI X -> Gradle current" and separate task for "Gradle X -> TAPI current". This allows to conveniently intercept clasloading of test classes in JUnit platform so that the right Tooling API under test can be loaded with the test.