Use a registrar for repositories declared in settings
This commit changes how the central dependencies declaration initialization
works. Instead of using the repositories from settings _or_ the repositories
from projects, we now use a registar in settings. This registar is used to
collect "rules" about repositories.

When a project is requesting repositories, we will now apply the rules to
the project's repository, based on the configuration in settings. That is
to say that each project will still use its repository handler, but the
rules will be applied independently to each of them. This should increase
the memory usage a bit compared to the previous solution, but it provides
better isolation, and allows querying for the repository list of a project
within a plugin.