Adjust expected error message
Due to #15817, we do not register global substations for the root build
anymore as long as it is not explicitly included.
In the case of this test, it leads to no substitution registered at all.
Because of that, we don't need to resolve during task graph calculation
anymore. So we fail later when the task starts.