Remove 'search-upwards', 'use-empty-settings', and 'master' dir support (#15879)
This removes the ability to...
- use a settings file in a 'master' directory outside the build
- ignore an existing settings file from a build script using 
  `StartParameter.useEmptySettings()`
- change the _search upwards_ behaviour from a build script using
  `StartParameter.setSearchUpwards()`
- change the _search upwards_ behaviour via the Tooling API connection
  (this is internal API). To keep the cross-version protocol intact,
  the `ProviderOperationParameters.isSearchUpwards()` method stays, 
  but its value has no effect when targeting a Gradle 7+ build.

While the functionality to ignore a settings file (e.g. when loading
from config cache) and to not search upwards (e.g. for buildSrc and
included builds) stay, everything only needed to make this configurable
from the launcher side is removed.

The Tooling API tests where using `connector.searchUpwards(false)` to
test Gradle projects without settings file. These are effectively
incomplete projects. So now, if we run the tests with Gradle 6+
Tooling API (as search upwards was deprecated in 6) we add an empty
settings file instead of `searchUpwards(false)` option.