Ignore buildSrc directory that does not contain valid Gradle build
Without this change, Gradle will execute a `buildSrc` build whenever a `buildSrc`
directory is detected, even if this directory contains no source files. This happens
frequently when migrating a build with `buildSrc` to one using composite builds:
the `buildSrc/build` directory which is normally ignored by git can be retained
resulting in unnecessary build invocation.

What's worse, the empty `buildSrc.jar` that is generated for an empty `buildSrc`
directory will be added to the build script classpath, which can trigger cache
misses for any custom tasks.

This fix addresses the issue by attempting to detect if `buildSrc` is indeed a valid
Gradle build, by checking for `settings.gradle[.kts]`, `build.gradle[.kts]` and for
a non-empty `src` directory. If none of these are found, the `buildSrc` directory
will be ignored, and the `buildSrc.jar` will not be added to the buildscript classpath.

Fixes #13491