Fix multi-test lifecycle for unrolled tests
The interceptors were incorrectly being invoked twice for the first iteration of a data-driven test with multiple executions (which manifested itself by invoking the JUnit4 rule twice for the first iteration of data driven test).
This change adds a test to assert the lifecycle of multi-versions tests is correct and fixes the underlying problem.