Fix tests using isolated home in no daemon mode
Previously the user.home was not properly propagated which was a problem
when a single use daemon had to be started anyway.

The change also reverts unnecessary test changes.

The root cause for the issue discovery is the migration to Spock 2 which
caused tests having a `where:` to be unrolled without the need for an
annotation.