Don't use exceptions for validation errors
Validation previously generated one exception per validation problem.
This causes a LOT of noise when using `--stacktrace` without bringing
any kind of user facing value. Exceptions shouldn't be used for
reporting "normal" problems: the user doesn't care where, in our
codebase, the problem is thrown. The only useful information to them
is the validation issue itself.

For us, we _might_ want to know where the validation error is thrown,
which is why in this commit we end up throwing a _single_ validation
exception, which, in turn, contains a rich description (using the
tree formatter).