Decide if current directory is the project root only where needed (#15987)
This restores most of what was already done in #15879,
but fixes the following cases that were broken:

- Running a 'buildSrc' build directly in the 'buildSrc' folder where
  no settings file exists in that folder.
- Doing the above through the Tooling API.
- Not reading a property file further up the folder hierarchy when
  running 'gradle init'.

Basically, what was missed is that the launcher also needs to know the
location of the root project to pick up the right 'gradle.properties'.

The  decision of whether the current project folder is used as root,
even if it does not contain a settings file, is now located in a single
place -- StartParameterInternal. useLocationAsProjectRoot().
This is used in all three places where this decision needs to be made:

- StartParameterInternal.isSearchUpwards()
  needed to determine the actual root (settings file)
- DeprecateUndefinedBuildWorkExecutor.execute()
  Check if one of the special cases apply or if there is a settings file at all
- LayoutToPropertiesConverter.configureFromBuildDir()
  Find the location of `gradle.properties` (used in Launcher and Tooling API) 

Another special case is an included build that runs as part of a parent. In this
cases, the included build is allowed to *not* have a settings file. To handle this
case, `StartParameterInternal.doNotSearchUpwards()` is still required.