Remove temporary ignoring of plugin validation
3 different projects were ignoring `ValidatePlugins` in their own
way. In general, this is just a _temporary_ fix which should be
removed once a compatible wrapper is used. The problem is that
they all configured the validation using `withType()` which caused
_all_ tasks of that type to be configured, including the external
plugin validation.

In practice, this caused extra work to be executed during the
smoke test, without any good reason. This should make the build
faster.

Should we need to ignore validation again, it would be better
to introduce a convention to do this, so that we can track
and remove.