Disable parallel execution for smoke test
External plugin validation applies to all projects, because not
all plugins are applied to all projects. However, in a lot of
cases, the plugins are the same, which makes some verification
tasks "from cache".

The test accidentally relies on the fact that we're going to
verify in a specific order, and therefore expects validation
problems to appear in a single project. To make this consistent,
we need to disable parallel execution.