Decouple severity from the cacheability state
The existing validation code conflates the severity of a problem with
the cacheability state. This commit decouples things by making the
severity only talk about severity, and having the cacheability state
carried as part of the problem payload.

The enum value is still used for all problems which do not use the
new builder mechanism. The plan is to finally get rid of this enum
value.