Fix plugin import spec not used for files
If a catalog was imported from a file, the logic to associate a node
in the resolved dependency graph to the actual dependency was broken.
The consequence is that we wouldn't associate the import spec, which
defines plugin excludes, to the actual catalog being imported, and
as such the import spec was ignored.

This commit fixes the problem by removing the optimization logic,
which was broken, and instead relies on creating one configuration
per catalog being imported.

Fixes #16079