Rename `dependencies.toml`
This commit renames the default `dependencies.toml` file to
`libs.versions.toml` instead. The reasoning is that:

1. this file doesn't really talk about dependencies, but about
a version catalog
2. the default version catalog is named "libs", which is now
reflected in the file name
3. the extension `.versions.toml` explains what to expect
from the contents of this file. This might be of interest for
IDEs which may use this extension file name to provide dedicated
support for version catalog files.
4. it's a compromise with verbosity (`.version.catalog.toml`)
5. it potentially allows auto-registration of files by listing
by extension

Note that this only applies to the default file. There's no
restriction on how you call the file in practice when importing
it. Similarly, for published components, the extension remains
".toml", for 2 reasons:

a. the Maven repository layout file mapping strategy disallows
complex extensions
b. the kind of file it corresponds to is already encoded in a
more strongly typed fashion thanks to the variant attributes,
so it's technically less fragile than an extension.

Fixes #15328