Add an optional metadata section to TOML file
This metadata section for now only supports a single key,
"format.version", which is there in case we want to support
a different file format in the future.

If the metadata is missing, Gradle assumes that this is the
current file format: this is a convenience such as that users
don't have to deal with this issue which is mostly there for
published catalogs.

Fixes #15329