Ignore Groovy generated methods for validation
Currently the heuristic is made as narrow as possible, that
it to say that we check if a method is annotated with `@Generated`
and that it contains a `$`.

It may be that there are other methods we should ignore which
don't contain a `$` in the name, however, we have to keep a
contraint like this because just checking for `@Generated` isn't
good: for example for properties, Groovy will generate methods
annotated with `@Generated` and those need to be validated.

Fixes #16117