Turn incorrect use of input annotation into an error
This commit also introduces a dummy annotation, `@ValidationProblem`, which
can be used in tests to trigger a validation problem. This was required because
some tests actually verify the behavior of Gradle when a validation problem
happens but that the problem is not fatal (execution continues). If we turn
all validation problems into errors, then we don't have any annotation to use
in tests.