Use unique settings class loader scope for each build in tree
Prior to this, all settings scripts shared the same class loader scopes. This caused unnecessary creation of class loaders and their classes across build invocations. Each settings script was vying for the same entry in the class loader cache, causing the load of a subsequent settings script with a different export loader to replace any existing. The net result was that all of the descendant loaders (buildSrc and the projects, i.e. almost everything) were new on each build.

The triggering condition is to have different settings scripts load different dependencies/plugins. This is actually typical when using the Gradle Enterprise plugin in a composite build, as it is typically only applied to the root build.

See https://github.com/gradle/gradle/pull/16351