Remove lock state for unlocked configurations
When resolving a configuration that is unlocked, the locking subsystem
is now informed and drops any existing lock state for the configuration.
This enables to clean up the content of the lock file when changing
which configurations are locked.

This process will only occur during lock writing, either for a complete
or partial update.

Fixes #16387