Update test-distribution plugin to version 2.0-rc-6
Compared to 2.0-rc-5, this version offers:
* remote executors support session concept

This means that the additional overhead we introduced
when creating many more partitions than before (in
order to be flexible with distributing work) is
reduced now. The test JVM on the remote side keeps
running and now executes multiple small partitions
in a single session.