Fix early cyclic include issue
When a build is included early (i.e. in pluginManagement {}) the
the scoped visibility requires us to walk down the tree of builds
which might lead to fullz configuring other builds. This in turn
can lead to an attempt to configure a build for which configuration
has already started. In these cases, starting the configuration anew
should just be skipped.

Which just means: You can't apply a pre-compiled script plugin
to the build that is defining it. The (indirect) include cycle could
still make sense and thus we skip instead of giving an error.