Fix precompiled plugin id deduction with .gradle substring in id
A precompiled Groovy plugin with file name `dev.gradlefoo.whatever.gradle` would be assigned a plugin id `dev`
and not the expected `dev.gradlefoo.whatever`.
There was a bug in how the plugin id is constructed from the filename where it took the substring before
the first occurrence of `.gradle` and used that as the plugin id. The expected behavior is to look for the
last occurrence of `.gradle` in precompiled plugin filename.