Improve the validation error message layout
Now that we have better error messages, we can improve their layout. In particular,
with the new system, we can give more context but for compatibility reasons, they
were displayed as a "block of text".

This commit changes the rendering so that we have richer output. For example, you'd
read:

```
> Task :myTask FAILED

FAILURE: Build failed with an exception.

* What went wrong:
A problem was found with the configuration of task ':myTask' (type 'TaskWithAbsentNestedInput').
  - Type 'TaskWithAbsentNestedInput' property 'nested' doesn't have a configured value.

    Reason: This property isn't marked as optional and no value has been configured.

    Possible solutions:
      1. Assign a value to 'nested'.
      2. Mark property 'nested' as optional.

    Please refer to https://docs.gradle.org/7.0-20210310230000+0000/userguide/validation_problems.html#value_not_set for more details about this problem.
```

or:

```
A problem was found with the configuration of task ':test' (type 'DefaultTask').
  - Type 'DefaultTask' property 'input' has unsupported value 'task ':dependencyTask''.

    Reason: Type 'DefaultTask' cannot be converted to a directory.

    Possible solutions:
      1. Use a String or CharSequence path, for example 'src/main/java' or '/usr/include'.
      2. Use a String or CharSequence URI, for example 'file:/usr/include'.
      3. Use a File instance.
      4. Use a Path instance.
      5. Use a Directory instance.
      6. Use a RegularFile instance.
      7. Use a URI or URL instance.
      8. Use a TextResource instance.

    Please refer to https://docs.gradle.org/7.0-20210310230000+0000/userguide/validation_problems.html#unsupported_notation for more details about this problem.
```

The good:

- error messages are now much easier to read
- error messages are more actionable
- the new test fixtures made it really easy to change the error message
- some tests which weren't using the new fixture were fixed

The bad:

- deprecation warnings are very hard to deal with in tests, because we don't detect multiline messages
properly (this is due to 2 problems: we perform validation at the end of the build AND we don't have markers
for those messages, we do plain string parsing)
- indentation changes in some messages because of nested exceptions, so a convenience was added to take this
into account. Again, because the fixtures we use are very picky on the layout and don't tolerate mismatches
in spacing.

The goal of this PR is NOT to improve the old fixtures. They should probably be changed to tolerate whatever
format wrt whitespaces. Testing the actual error message should belong to unit testing, if we ever want to
do this.