Support Zulu-like Java distributions
When starting a Gradle daemon, we check that the Java home of the
just started daemon matches what was desired. This can be confused
by Java distributions like Zulu that structure the obvious "Java Home"
directory as several symlinks to another directory.

We end up deciding that the final target of the symlink is the real
Java Home. When we compare this to the desired Java Home by just
looking at the file path strings, these do not match and we fail
the build.

To workaround this, you can figure out the symlink yourself and
use a different (more hidden) path inside the Java distribution.

This may not fix all problems with comparing Java Home paths, but
this at least allows you to use Zulu to start a daemon.

Fixes #12840