Fix performance of test project generator
Otherwise it is too slow to generate a stress test for the incremental compiler.
Migrated all performance-critical sections to statically compiled Groovy.
Reworked the DependencyTree algorithm, so that inserting a single class only takes
log(c) time instead of c*p*log(c*p), where p is the number of projects and c is the
number of classes per project.

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>