Add classpath analysis to PreviusCompilationData
This way all the necessary data for incremental compilation is stored as a task
output instead of being put into the build cache as a side effect of classpath
snapshotting. This also means that the language-java project does not depend on
the build cache or execution engine directly and is thus a better example for
community plugins seeking to implement incremental tasks. Storing the merged
analysis also means that no merging is necessary on the next compilation.

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>