Fix compilation avoidance test
This test became dependent on the order the test methods are executed because some of the tests
care about the compile avoidance warnings while others don't. The tests that do care about
compile avoidance warnings only care about the warnings originating from user code (build scripts).
However, there are also compile avoidance warnings originating from library jars and those are
only output during the first test executed (as the libraries don't change between executions).
Given the first test executed does not care about the warnings, the subsequent tests executed
will only see the ones from user code and all is green. If a test that cares about warnings
is executed in isolation (or first in order), it will see all the warnings and it will fail
because of too many warnings.

This change filters out the uninteresting warnings so that the order of execution does not play a role.